refactor: extract browser-safe core from graphql-codegen into graphql-query
Move the pure-function query generation core (generators, AST builders,
naming helpers, introspection utilities, client layer, types) from
graphql-codegen into graphql-query as the canonical source.

graphql-codegen now depends on graphql-query and re-exports all moved
modules via thin shims for full backward compatibility.

Modules moved to graphql-query:
- generators/ (buildSelect, buildFindOne, buildCount, mutations, field-selector, naming-helpers)
- core/ast.ts, core/custom-ast.ts, core/query-builder.ts
- client/ (TypedDocumentString, error handling, execute)
- types/ (CleanTable, CleanField, QueryOptions, introspection, etc.)
- introspect/ (infer-tables, transform, transform-schema, schema-query)
- meta-object/ (convert, validate)
- utils (stripSmartComments)

Kept in codegen (Node.js-only):
- CLI entry points, file output, watch mode, config loading
- Babel codegen templates, React Query hook generation
- Database connection, schema source fetching

Both packages build clean, all 301 codegen tests + 18 query tests pass.

Related: https://github.com/constructive-io/constructive-planning/issues/651
