+++ graphql/query/src/introspect/index.ts
@@ -1,0 +1,10 @@
+/**
+ * Introspection barrel export
+ *
+ * Re-exports introspection utilities for converting GraphQL schemas to CleanTable format.
+ */
+
+export * from './infer-tables';
+export * from './transform';
+export * from './transform-schema';
+export * from './schema-query';
