+++ graphql/query/src/meta-object/convert.ts
@@ -1,6 +1,6 @@
-import type { MetaFieldType } from '../types';
+import type { MetaFieldType } from '../types';  // resolves to types/index.ts
 
-// Input schema types
+// Input schema types (from _meta query response)
 interface MetaSchemaField {
   name: string;
   type: MetaFieldType;
@@ -40,7 +40,7 @@ interface MetaSchemaInput {
   };
 }
 
-// Output types
+// Output types (internal MetaObject format)
 interface ConvertedField {
   name: string;
   type: MetaFieldType;
@@ -71,8 +71,11 @@ interface ConvertedMetaObject {
   tables: ConvertedTable[];
 }
 
+/**
+ * Convert from raw _meta schema response to internal MetaObject format
+ */
 export function convertFromMetaSchema(
-  metaSchema: MetaSchemaInput
+  metaSchema: MetaSchemaInput,
 ): ConvertedMetaObject {
   const {
     _meta: { tables },
@@ -90,7 +93,7 @@ export function convertFromMetaSchema(
       uniqueConstraints: pickArrayConstraint(table.uniqueConstraints),
       foreignConstraints: pickForeignConstraint(
         table.foreignKeyConstraints,
-        table.relations
+        table.relations,
       ),
     });
   }
@@ -99,7 +102,7 @@ export function convertFromMetaSchema(
 }
 
 function pickArrayConstraint(
-  constraints: MetaSchemaConstraint[]
+  constraints: MetaSchemaConstraint[],
 ): ConvertedConstraint[] {
   if (constraints.length === 0) return [];
   const c = constraints[0];
@@ -108,7 +111,7 @@ function pickArrayConstraint(
 
 function pickForeignConstraint(
   constraints: MetaSchemaForeignConstraint[],
-  relations: MetaSchemaRelations
+  relations: MetaSchemaRelations,
 ): ConvertedForeignConstraint[] {
   if (constraints.length === 0) return [];
