+++ graphql/query/src/meta-object/validate.ts
@@ -1,15 +1,39 @@
-import Ajv from 'ajv';
+import Ajv, { type ValidateFunction } from 'ajv';
 
 import format from './format.json';
 
-export function validateMetaObject(obj: any) {
-  const ajv = new Ajv({ allErrors: true });
-  const valid = ajv.validate(format, obj);
+export interface ValidationResult {
+  errors: unknown[] | null | undefined;
+  message: string;
+}
+
+let cachedAjv: Ajv | null = null;
+let cachedValidator: ValidateFunction | null = null;
+
+function getValidator() {
+  if (!cachedAjv) {
+    cachedAjv = new Ajv({ allErrors: true });
+    cachedValidator = cachedAjv.compile(format);
+  }
+
+  return {
+    ajv: cachedAjv,
+    validator: cachedValidator!,
+  };
+}
+
+/**
+ * Validate a MetaObject against the JSON schema
+ * @returns true if valid, or an object with errors and message if invalid
+ */
+export function validateMetaObject(obj: unknown): true | ValidationResult {
+  const { ajv, validator } = getValidator();
+  const valid = validator(obj);
 
   if (valid) return true;
 
   return {
-    errors: ajv.errors,
-    message: ajv.errorsText(ajv.errors, { separator: '\n' }),
+    errors: validator.errors,
+    message: ajv.errorsText(validator.errors, { separator: '\n' }),
   };
 }
