+++ graphql/query/src/types/index.ts
@@ -1,0 +1,24 @@
+/**
+ * Barrel export for all types
+ *
+ * Re-exports both the original QueryBuilder runtime types (core.ts)
+ * and the codegen-style types (schema, introspection, query, mutation, selection)
+ */
+
+// Original QueryBuilder runtime types
+export * from './core';
+
+// Codegen-style schema types (CleanTable, CleanField, etc.)
+export * from './schema';
+
+// Introspection types
+export * from './introspection';
+
+// Query types (QueryOptions, Filter, etc.)
+export * from './query';
+
+// Mutation types
+export * from './mutation';
+
+// Selection types (FieldSelection presets, SimpleFieldSelection, etc.)
+export * from './selection';
